[FIX] web: correctly display header tooltips in list
A traceback occured when displaying the tooltip (only visible in debug) on a list
view header if the field had a `widget` attribute (introduced in rev. odoo/odoo@5dc971b).

The issue here was linked to the fact that a widget is not always rendered in list view
(compared to form view) so the `Widget` property in the attributes was not there.

Also, the widget description was displayed in the list header (instead of
the field name).

Task 1981757

closes odoo/odoo#33060

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>